conn: add serialization for tuples (and change ser. interfaces as a consequence)
tuple is not standart_layout type, so, separate procedure is required.
This procedure is recursive to allow nested tuples (not tested yet).
Because of it also now user-provided storage (buffer) is required
Earlier the same memory (the value) was used, just reinterpreted;
now it's not possible because memory layout of serialized tuple
(determined by me) is different from that of the tuple itself.