dsl: make Callable-s copyable; make ECall be able to store (own) it's Callable; add LogFun
Needed for the new LogFun DSL function.
LogFun can be seen as prototype for Resident functions that can be called only in DSL code.
ECall now loses (as it seems) type inference on construction (similar to std::forward)
but can be used for both references to Callable (i.e. as in make_generic_dsl_fun)
and temporary Callable-s (as for the new generic logging function LogFun).